Allow for smooth upgrades

+++ deploy_platform.sh
@@ -25,31 +25,24 @@ if [ -z "$RF_QUAY_USERNAME" ] || [ -z "$RF_QUAY_PASSWORD" ]; then
   exit 1
 fi
 
+print_bgreen "Logging in to Helm registry"
 echo -n ${RF_QUAY_PASSWORD} | helm registry login -u ${RF_QUAY_USERNAME} --password-stdin ${QUAY_DOCKER_REGISTRY}
-# Function to pause and wait for user input
-# pause(){
-#     read -p "Did you complete quay.io registry login? (y/n): " choice
-#     case "$choice" in
-#         y|Y ) print_bgreen "Continuing...";;
-#         * ) print_bred "Exiting..."; exit 1;;
-#     esac
-# }
 
-# pause
+if ! kind get clusters | grep rapidfort-platform > /dev/null; then
+  print_bgreen "Starting kind"
+  kind create cluster --name rapidfort-platform --config kind-config.yaml --wait 8m
 
-print_bgreen "Starting kind"
-kind create cluster --name rapidfort-platform --config kind-config.yaml --wait 8m
+  print_bgreen "Ingress start"
+  kubectl apply -f ingress-nginx.yaml
 
-print_bgreen "Ingress start"
-kubectl apply -f ingress-nginx.yaml
+  kubectl wait --namespace ingress-nginx \
+    --for=condition=ready pod \
+    --selector=app.kubernetes.io/component=controller \
+    --timeout=200s
 
-kubectl wait --namespace ingress-nginx \
-  --for=condition=ready pod \
-  --selector=app.kubernetes.io/component=controller \
-  --timeout=200s
-
-print_bgreen "Remove admission controller"
-kubectl delete -A ValidatingWebhookConfiguration ingress-nginx-admission
+  print_bgreen "Remove admission controller"
+  kubectl delete -A ValidatingWebhookConfiguration ingress-nginx-admission
+fi
 
 print_bgreen "Install Image Pull Secret"
 kubectl apply -f secret.yaml
